Add basic assignment

+++ examples/execve2.py
@@ -22,8 +22,8 @@ def hello(ctx: c_void_p) -> c_int32:
 @section("tracepoint/syscalls/sys_exit_execve")
 def hello_again(ctx: c_void_p) -> c_int64:
     print("exited")
-    key = c_int64(0)
-    tsp = last.lookup(key)
+    key = 0
+    tsp = last().lookup(key)
     print(tsp)
     ts = bpf_ktime_get_ns()
     return c_int64(0)
@@ -32,5 +32,6 @@ def hello_again(ctx: c_void_p) -> c_int64:
 # @bpfglobal
 # def LICENSE() -> str:
 #     return "GPL"
+
 
 LICENSE = "GPL"
